+++ Dijkstras/Classes/NodeConnection.cs
@@ -3,9 +3,9 @@
     public class NodeConnection
     {
         public Node Node { get; private set; }
-        public int Distance { get; private set; }
+        public double Distance { get; private set; }
 
-        public NodeConnection(Node targetNode, int distance)
+        public NodeConnection(Node targetNode, double distance)
         {
             Node = targetNode;
             Distance = distance;
